ada_tools/requirements.txt
# This file is autogenerated by pip-compile
azure-storage-blob>=12.8.1
affine>=2.3.0             # via ada_tools (setup.py), rasterio
attrs>=19.3.0             # via ada_tools (setup.py), fiona, rasterio
beautifulsoup4>=4.9.1     # via ada_tools (setup.py), bs4
bs4>=0.0.1                # via ada_tools (setup.py)
click-plugins>=1.1.1      # via ada_tools (setup.py), fiona, rasterio
click>=7.1.2              # via ada_tools (setup.py), click-plugins, cligj, fiona, rasterio
cligj>=0.5.0              # via ada_tools (setup.py), fiona, rasterio
fiona>=1.8.13.post1       # via ada_tools (setup.py), geopandas
# gdal>==3.1.2               # via ada_tools (setup.py)  # installed separately
geopandas>=0.10.0          # via ada_tools (setup.py)
munch>=2.5.0              # via ada_tools (setup.py), fiona
numpy>=1.23.5             # via ada_tools (setup.py), pandas, rasterio, snuggs
Pillow==10.0.0
pandas>=1.3.1             # via ada_tools (setup.py), geopandas
pygeos>=0.8.0             # added manually
pyparsing>=2.4.7          # via ada_tools (setup.py), snuggs
pyproj>=2.6.1.post1       # via ada_tools (setup.py), geopandas
python-dateutil>=2.8.1    # via ada_tools (setup.py), pandas
pytest-mock>=3.11.1
pytz>=2020.1              # via ada_tools (setup.py), pandas
rasterio>=1.2.10          # via ada_tools (setup.py)
scipy>=1.9.3              # added manually
shapely>=1.8.4            # via ada_tools (setup.py), geopandas
six>=1.15.0               # via ada_tools (setup.py), fiona, munch, python-dateutil
snuggs>=1.4.7             # via ada_tools (setup.py), rasterio
soupsieve>=2.0.1          # via ada_tools (setup.py), beautifulsoup4
tqdm>=4.48.2              # via ada_tools (setup.py)
dateparser>=0.7.6
matplotlib==3.6.3
scikit-image==0.19.3
scipy==1.9.3
